Removing old, commented extension code
Since we no longer need the user to supply info about the file extensions used as we grab this automatically from the file on disk, the old code that was commented is now removed.

Also removed the 'Remove MKV' option as I do not feel it is needed. Current functionality simply maintains the original recording file extension.