Add notes about new python module prereqs
With the addition of pywinctl as our module of choice for grabbing foreground window names, this is the first additional module the user will need to install.

Initially just the prereqs for Windows are in. Other OSes may have additional modules needed. Further testing or investigation is needed to confirm these.